eat default-headers
