+++ packages/insomnia-inso/package.json
@@ -63,7 +63,6 @@
     "cosmiconfig": "^6.0.0",
     "enquirer": "^2.3.6",
     "insomnia-plugin-core-themes": "^3.19.0-beta.4",
-    "insomnia-plugin-default-headers": "^3.19.0-beta.4",
     "insomnia-send-request": "^3.19.0-beta.4",
     "insomnia-testing": "^3.19.0-beta.4",
     "nedb": "^1.8.0",
