--- plugins/insomnia-plugin-default-headers/index.js
@@ -1,1 +1,0 @@
-module.exports.requestHooks = [require('./src/default-headers')];
