--- plugins/insomnia-plugin-default-headers/package.json
@@ -1,21 +1,0 @@
-{
-  "private": true,
-  "name": "insomnia-plugin-default-headers",
-  "version": "3.19.0-beta.4",
-  "author": "Kong <office@konghq.com>",
-  "description": "Various data importers for Insomnia",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Kong/insomnia.git",
-    "directory": "plugins/insomnia-plugin-default-headers"
-  },
-  "bugs": {
-    "url": "https://github.com/Kong/insomnia/issues"
-  },
-  "main": "index.js",
-  "insomnia": {
-    "name": "default-headers",
-    "description": "Add default headers to requests"
-  }
-}
